+++ _common/gameModel/libs/ClientScript/gameMap/main.ts
@@ -1,7 +1,7 @@
 import { LOCATION_ENUM } from '../game/common/simulationState/locationState';
 import { getTypedInterfaceState } from '../gameInterface/interfaceState';
 import { bringOverlayItemToFront, toggleOverlayItem } from '../gameMap/mapEntities';
-import { Point } from '../map/point2D';
+import { Point } from '../tools/point2D';
 import {
   getAvailableActionTemplateById,
   isChoiceTemplate,
