+++ _common/gameModel/libs/ClientScript/tools/point2D.ts
